test: add integration tests for API and DNS operations
