+++ tests/api/mod.rs
@@ -1,0 +1,4 @@
+mod basic;
+mod dns;
+mod record;
+mod zone;
