+++ tests/lib.rs
@@ -1,0 +1,2 @@
+mod api;
+mod common;
